fixed bug in slideCharWindow and nullByteTransitions
