+++ src/jsMain/kotlin/FileUpload.kt
@@ -1,0 +1,56 @@
+import org.khronos.webgl.*
+import org.w3c.files.File
+import org.w3c.files.FileReader
+import kotlinx.browser.document
+
+fun arrayBufferToHex(buffer: ArrayBuffer): String {
+    val byteArray = Uint8Array(buffer) // Create a Uint8Array view for the buffer
+    val dynamic = byteArray.asDynamic()
+    return (0 until byteArray.length).joinToString("") { index ->
+        val b16string = dynamic[index].toString(16) as String
+        b16string.padStart(2, '0')
+    }
+}
+
+// read binary file and add content to textarea
+fun readBinaryFile(file: File) {
+    val reader = FileReader()
+
+    reader.onload = {
+        val arrayBuffer = reader.result as? ArrayBuffer
+        if (arrayBuffer != null) {
+            val hexContent = arrayBufferToHex(arrayBuffer) // Convert binary data to hex
+            // Display hex content in the textarea
+            appendTextareaWithContent(hexContent)
+        } else {
+            console.error("Failed to read binary file content")
+        }
+    }
+
+    reader.onerror = {
+        console.error("Failed to read the file: ${reader.error?.message}")
+    }
+
+    reader.readAsArrayBuffer(file) // Read binary data in the file
+}
+
+// read txt file and append to textarea
+fun readFile(file: File) {
+    val reader = FileReader()
+
+    reader.onload = {
+        val content = reader.result?.toString() // Safely convert `result` to a string
+        if (content != null) {
+            // Write the file content to the textarea
+            appendTextareaWithContent(content)
+        } else {
+            console.error("File content is null")
+        }
+    }
+
+    reader.onerror = {
+        console.error("Failed to read the file: ${reader.error?.message}")
+    }
+
+    reader.readAsText(file) // Read the file content as text
+}
