+++ src/jsMain/resources/index.html
@@ -224,6 +224,13 @@
             border: 1px solid #ccc;
             background: white;
             vertical-align: middle;
+            text-align: center;
+        }
+
+        .ascii-char {
+            display: block;
+            font-size: 10px;
+            color: #888;
         }
 
         .highlightbyte {
@@ -470,7 +477,7 @@
 
         .separator-placeholder {
             display: inline-block;
-            height: 22px;
+            height:38px;
             background-color: transparent;
             cursor: pointer;
             vertical-align: middle;
